# app/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Noeko Developer API
  version: 0.1.0
  summary: User-scoped automation API for Noeko
  description: |
    The implemented Noeko API for personal scripts, agents, and integrations.
    Personal API credentials use explicit scopes. Browser session bearer tokens may
    also call authenticated endpoints and are required to manage API credentials.
servers:
  - url: /api/v1
    description: Current Noeko installation
...
security:
  - apiCredential: []
  - browserSession: []
paths:
...
  /ideas:
    get:
      operationId: listIdeas
      summary: List accessible ideas
      tags: [Ideas]
      x-required-scope: ideas:read
      parameters:
        - name: limit
          in: query
          description: Number of ideas to return. Values are clamped to 1 through 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: start
          in: query
          description: Zero-based offset. This is an early contract and may be replaced by cursors.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Ideas ordered by most recently updated
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdeaListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createIdea
      summary: Create a private idea
      description: Embeddings and other derived data are processed asynchronously.
      tags: [Ideas]
      x-required-scope: ideas:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaInput"
      responses:
        "201":
          description: Idea created
          headers:
            Location:
              description: Relative URL of the created idea
              schema:
                type: string
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdeaResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    apiCredential:
      type: http
      scheme: bearer
      bearerFormat: Noeko API credential
      description: Personal credential beginning with `noeko_live_`.
    browserSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Browser access token. Required for credential-management endpoints.
  headers:
    RequestId:
      description: Request correlation ID, generated unless supplied by the client.
      schema:
        type: string
...
  responses:
    InvalidRequest:
      description: Request input is invalid
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthenticated:
      description: Bearer token is absent or invalid
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Account is disabled or credential lacks the required scope
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Operation failed unexpectedly
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    IdeaInput:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        content:
          type: string
          default: ""
          description: Rich HTML stored by the current editor implementation.
...
    Idea:
      type: object
      required:
        - id
        - object
        - title
        - content
        - content_format
        - visibility
        - created_at
        - updated_at
        - content_updated_at
        - processing
      properties:
        id:
          type: string
        object:
          type: string
          const: idea
        title:
          type: string
        content:
          type: string
        content_format:
          type: string
          const: html
        visibility:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        content_updated_at:
          type: string
          format: date-time
        processing:
          type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [pending, completed]
    IdeaResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Idea"
    IdeaListResponse:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Idea"
        page:
          type: object
          required: [has_more]
          properties:
            has_more:
              type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code]
          properties:
            type:
              type: string
              enum:
                - authentication_error
                - permission_error
                - not_found
                - invalid_request
                - internal_error
            code:
              type: string
            message:
              type: string